move time reducers
